--- lib/issueCommentHandler.js
@@ -1,60 +1,0 @@
-'use strict';
-
-const github = require('./github');
-const bumpVersion = require('semver-increment');
-const promisify = require('es6-promisify');
-const getMasksFromCommentBody = require('./function').getMasksFromCommentBody;
-const getPrMsgFormUrl = require('./function').getPrMsgFormUrl;
-const getBranchMsgFormHtmlUrl = require('./function').getBranchMsgFormHtmlUrl;
-
-const getPullRequests = promisify(github.pullRequests.get.bind(github.pullRequests));
-const getContent = promisify(github.repos.getContent.bind(github.repos));
-const updateFile = promisify(github.repos.updateFile.bind(github.repos));
-
-function issueCommentHandler(event) {
-  const payload = event.payload;
-  const pullRequests = payload.issue.pull_request;
-  if (!pullRequests) {
-    console.log('It isn\'t a pull request.');
-    return;
-  }
-  const commentBody = payload.comment.body.toLowerCase();
-  const masks = getMasksFromCommentBody(commentBody);
-  if (!commentBody.match('@seanbot') || !commentBody.match('bump' || !masks)) {
-    console.log(commentBody, 'It\'s not a bump action.');
-    return;
-  }
-  getPullRequests(getPrMsgFormUrl(pullRequests.url))
-  .then((res) => {
-    const baseMsg = {
-      user: res.base.repo.owner.login,
-      repo: res.base.repo.name,
-      path: 'package.json',
-      ref: res.base.ref,
-    };
-    const headMsg = {
-      user: res.head.repo.owner.login,
-      repo: res.head.repo.name,
-      path: 'package.json',
-      ref: res.head.ref,
-    };
-    return Promise.all([getContent(baseMsg), getContent(headMsg)]);
-  })
-  .then((results) => {
-    const baseRes = results[0];
-    const headRes = results[1];
-    const basePackageJson = JSON.parse(new Buffer(baseRes.content, 'base64').toString());
-    const headPackageJson = JSON.parse(new Buffer(headRes.content, 'base64').toString());
-    headPackageJson.version = bumpVersion(masks, basePackageJson.version);
-    const msg = getBranchMsgFormHtmlUrl(headRes.html_url);
-    msg.path = 'package.json';
-    msg.message = `version: ${basePackageJson.version} bump to :version ${headPackageJson.version}`;
-    msg.content = new Buffer(JSON.stringify(headPackageJson, null, '  ')).toString('base64');
-    msg.sha = headRes.sha;
-    return updateFile(msg);
-  })
-  .then((res) => console.log(res))
-  .catch((err) => console.error(err.message && err.stack));
-}
-
-module.exports = issueCommentHandler;
